requirements.txt
streamlit
langchain
langchain-community
langchain-google-genai
faiss-cpu
pandas
python-dotenv
pypdf
tiktoken
unstructured
markdown
langchain-openai
